refactor: 패키지명 controller에서 application으로 변경
* controller에 다양한 interface가 있어 의미 불명확
  * application으로 변경하며 앱 계층의 필수 요소를 표현
